Bugfix: Buidsteps cannot be run from database container
When initializing a project the default.sql.gz is imported. Because
these file is likely not fully up to date some aditional build steps
like drush config:import must be run. PHP and drush are not available in
the database container so the contiduional_import.sh script must be
started in the appserver container.